+++ HISTORY.rst
@@ -5,6 +5,9 @@ Next Release
 ------------
 * Add test ``find_medium_metabolites`` to detect and display all substrates
   capable of being taken up by model
+* Display the score on History and Snapshot reports.
+* Invert the colour coding.
+* Display results as percentages.
 
 0.7.6 (2018-05-28)
 ------------------